fix: CO2concentration moved to database object, fixed error in dutchsoilprofiles
The CO2concentration file needed to start at the starting year of the simulation. Therefore I could not simply copy the Atmospheric.co2 file, but needed to modify it first. This is now handled by the CO2concentration database. Additionally, the get_swapinput_cofani method from SoilProfile in dutchsoilprofiles had the wrong length.