fix: Tests are passing again
Issue was because the validation.yaml was changed, a cftb was not allowed when swcf=2, which was the case when importing old swap files for grass for testing. Solution was to have CFTB only have CF, and CHTB CH and CF, so reading the old SWAP files would return a chtb instead of a cftb. Then the validation doesnt give an error anymore.