break: swap executables are not shipped with pyswap anymore, use github instead
In order to be in line with licences and enable more flexibility, the executables are not shipped with the package anymore via pypi and are instead collected from an unofficial repository of SWAP fortran code. The user instead can use the CLI or the psp.get_swap() function to obtain the copy of the executable.

Additionally version tracking was improved.