perf: stop copying the executable to tempdirs
Implemented a feature improving performance where the executables are not copied to the tempdirs anymore but each command (including those in parallel) is run from the cwd in the tempdir and references the executable in the main libs. This worked also for multiprocessing.

Additionall, SWAP executable is now verified upon running the model. If it's not present in the libs directory, it will be automatically downloaded by get-swap utility.

closes #107